+++ RELEASE-NOTES.md
@@ -20,6 +20,8 @@
 
 ## New additions
 * Added `--retain-comments` option to `snow sql` command to allow passing comments to Snowflake.
+* Added `--replace` and `--if-not-exists` options to `snow object create` command.
+
 ## Fixes and improvements
 * `snow --info` callback returns information about `SNOWFLAKE_HOME` variable.
 * Removed requirement of existence of any `requirements.txt` file for Python code execution via `snow git execute` command.